requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
playwright>=1.40.0
pytest>=7.0.0
